slight changes to OpDsl: - add missing CokoMarker annotation - inline group function - move Op.getAllNodes + getNodes into other package
