+++ codyze-v3/codyze-specification-languages/coko/coko-core/src/main/kotlin/modelling/OpComponents.kt
@@ -5,6 +5,7 @@ import de.fraunhofer.aisec.codyze.specification_languages.coko.coko_core.CokoMar
 /** Represents a parameter in the [Signature] */
 typealias Parameter = Any?
 
+@CokoMarker
 /** Represents a group of parameters that all belong to the same index */
 class ParameterGroup {
     val parameters = arrayListOf<Parameter>()
