--- codyze-v3/codyze-specification-languages/coko/coko-core/src/main/kotlin/ordering/OrderEvaluator.kt
@@ -1,3 +1,0 @@
-package de.fraunhofer.aisec.codyze.specification_languages.coko.coko_core.ordering
-
-class OrderEvaluator {}
